add a common object for browse and search request
